requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-filter==25.1
django-jazzmin==3.0.1
django-modeltranslation==0.19.15
django-rosetta==0.10.2
djangorestframework==3.16.0
Faker==37.4.0
idna==3.10
Markdown==3.8.2
pillow==11.2.1
polib==1.2.0
psycopg2-binary==2.9.10
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
